+++ android-app/src/main/java/com/ricoh/livestreaming/app/Config.kt
@@ -5,6 +5,7 @@
 package com.ricoh.livestreaming.app
 
 import android.content.Context
+import com.ricoh.livestreaming.IceServersProtocol
 import org.webrtc.Logging
 
 /**
@@ -17,6 +18,7 @@ object Config {
     private const val KEY_ROOM_ID = "room_id"
     private const val KEY_LOGGING_SEVERITY = "logging_severity"
     private const val KEY_ROOM_TYPE = "room_type"
+    private const val KEY_ICE_SERVERS_PROTOCOL = "ice_servers_protocol"
 
     enum class LoggingSeverity(val level: String, val position: Int, val severity: Logging.Severity) {
         VERBOSE("VERBOSE", 0, Logging.Severity.LS_VERBOSE),
@@ -34,12 +36,21 @@ object Config {
         P2P_TURN(2, R.id.room_type_p2p_turn, RoomSpec.RoomType.P2P_TURN),
     }
 
+    enum class IceServersProtocol(val position: Int, val id: Int, val type: com.ricoh.livestreaming.IceServersProtocol) {
+        ALL(0, R.id.ice_servers_protocol_all, com.ricoh.livestreaming.IceServersProtocol.ALL),
+        UDP(1, R.id.ice_servers_protocol_udp, com.ricoh.livestreaming.IceServersProtocol.UDP),
+        TCP(2, R.id.ice_servers_protocol_tcp, com.ricoh.livestreaming.IceServersProtocol.TCP),
+        TLS(3, R.id.ice_servers_protocol_tls, com.ricoh.livestreaming.IceServersProtocol.TLS),
+    }
+
     private val DEFAULT_ROOM_TYPE = RoomType.SFU
+    private val DEFAULT_ICE_SERVERS_PROTOCOL = IceServersProtocol.ALL
 
     /** Configuration values    */
     private var roomId = ""
     private var loggingSeverity = "INFO"
     private var roomType = RoomType.SFU.position
+    private var iceServersProtocol = DEFAULT_ICE_SERVERS_PROTOCOL.position
 
     fun load(context: Context) {
         val preferences = context.getSharedPreferences(CONFIG_FILE, Context.MODE_PRIVATE)
@@ -48,6 +59,7 @@ object Config {
         this.loggingSeverity = preferences.getString(KEY_LOGGING_SEVERITY, "").toString()
         if (this.loggingSeverity.isEmpty()) this.setLoggingSeverity(context, "INFO")
         this.roomType = preferences.getInt(KEY_ROOM_TYPE, DEFAULT_ROOM_TYPE.position)
+        this.iceServersProtocol = preferences.getInt(KEY_ICE_SERVERS_PROTOCOL, DEFAULT_ICE_SERVERS_PROTOCOL.position)
     }
 
     fun getRoomId(): String {
@@ -112,6 +124,37 @@ object Config {
             else -> RoomType.SFU.id
         }
         this.saveInt(context, KEY_ROOM_TYPE, roomType)
+    }
+
+    fun getSelectedIceServersProtocolID(): Int {
+        return when (this.iceServersProtocol) {
+            IceServersProtocol.ALL.position -> IceServersProtocol.ALL.id
+            IceServersProtocol.UDP.position -> IceServersProtocol.UDP.id
+            IceServersProtocol.TCP.position -> IceServersProtocol.TCP.id
+            IceServersProtocol.TLS.position -> IceServersProtocol.TLS.id
+            else -> this.DEFAULT_ICE_SERVERS_PROTOCOL.id
+        }
+    }
+
+    fun getIceServersProtocol(): com.ricoh.livestreaming.IceServersProtocol {
+        return when (this.iceServersProtocol) {
+            IceServersProtocol.ALL.position -> com.ricoh.livestreaming.IceServersProtocol.ALL
+            IceServersProtocol.UDP.position -> com.ricoh.livestreaming.IceServersProtocol.UDP
+            IceServersProtocol.TCP.position -> com.ricoh.livestreaming.IceServersProtocol.TCP
+            IceServersProtocol.TLS.position -> com.ricoh.livestreaming.IceServersProtocol.TLS
+            else -> this.DEFAULT_ICE_SERVERS_PROTOCOL.type
+        }
+    }
+
+    fun setIceServersProtocol(context: Context, id: Int) {
+        this.iceServersProtocol = when (id) {
+            IceServersProtocol.ALL.id -> IceServersProtocol.ALL.position
+            IceServersProtocol.UDP.id -> IceServersProtocol.UDP.position
+            IceServersProtocol.TCP.id -> IceServersProtocol.TCP.position
+            IceServersProtocol.TLS.id -> IceServersProtocol.TLS.position
+            else -> IceServersProtocol.ALL.id
+        }
+        this.saveInt(context, KEY_ICE_SERVERS_PROTOCOL, iceServersProtocol)
     }
 
     private fun saveString(context: Context, key: String, value: String) {