+++ android-app/src/main/java/com/ricoh/livestreaming/app/RecvActivity.kt
@@ -92,6 +92,7 @@ class RecvActivity : AppCompatActivity() {
                                 .sendingPriority(SendingVideoOption.SendingPriority.NORMAL)
                                 .maxBitrateKbps(BuildConfig.VIDEO_BITRATE)
                                 .build()))
+                .iceServersProtocol(Config.getIceServersProtocol())
                 .build()
 
         mClient!!.connect(